requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asgiref==3.11.1
bcrypt==3.1.7
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
greenlet==3.2.5
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pyasn1==0.6.2
pycparser==2.23
PyJWT==2.11.0
python-jose==3.5.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
